+++ test/libsolidity/syntaxTests/functionCalls/calloptions_repeated.sol
@@ -4,7 +4,6 @@ contract C {
         this.foo{value:2, gas: 5}{gas:2};
         (this.foo{value:2, gas: 5}){gas:2};
         this.foo{value:2, gas: 5}{value:6};
-        this.foo.value(4){value:2, gas: 5};
         this.foo{gas:2, value: 5}{value:2, gas:5};
         new D{salt:"abc"}{salt:"a"}();
     }
@@ -15,8 +14,6 @@ contract C {
 // TypeError: (78-110): Option "gas" has already been set.
 // TypeError: (120-154): Option "gas" has already been set.
 // TypeError: (164-198): Option "value" has already been set.
-// Warning: (208-222): Using ".value(...)" is deprecated. Use "{value: ...}" instead.
-// TypeError: (208-242): Option "value" has already been set.
-// TypeError: (252-293): Option "value" has already been set.
-// TypeError: (252-293): Option "gas" has already been set.
-// TypeError: (303-330): Option "salt" has already been set.
+// TypeError: (208-249): Option "value" has already been set.
+// TypeError: (208-249): Option "gas" has already been set.
+// TypeError: (259-286): Option "salt" has already been set.